Photon firmware flash of blink.ino

+++ photon/blink.ino
@@ -1,6 +1,6 @@
 int led1 = D0; // Instead of writing D0 over and over again, we'll write led1
 int led2 = D7; // Instead of writing D7 over and over again, we'll write led2
-int delay = 200;
+unsigned long delay = 200;
 
 void setup() {
   pinMode(led1, OUTPUT);
